+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsContributionProvidersBoxTestGenerated.java
@@ -1956,6 +1956,12 @@ public class JsContributionProvidersBoxTestGenerated extends AbstractJsContribut
       }
 
       @Test
+      @TestMetadata("AssistedSuspendFunctionParameter.kt")
+      public void testAssistedSuspendFunctionParameter() {
+        run("AssistedSuspendFunctionParameter.kt");
+      }
+
+      @Test
       @TestMetadata("AssistedTypesCanBeExplicitlyProvided.kt")
       public void testAssistedTypesCanBeExplicitlyProvided() {
         run("AssistedTypesCanBeExplicitlyProvided.kt");